所有的controler的用户验证都添加了

+++ controler/refund.php
@@ -8,6 +8,17 @@
  */
 include_once('../config/config.php');
 include_once('../'.$libs_dir.'/db.php');
+require_once("../models/config.php");
+
+$user_id = null;
+if(isset($loggedInUser) && $loggedInUser->user_id){
+    $user_id = $loggedInUser->user_id;
+}
+if(!$user_id){
+    $result = array("bizCode" => 0, "memo" => "用户未登", "data"=>array("status" => 100));
+    echo json_encode($result);
+    exit;
+}
 
 $db = new DB($db_name,$db_host,$db_username,$db_password);
 $db->query("SET NAMES 'UTF8'");
@@ -19,7 +30,7 @@ if(!$id || !$details){
     $result['bizCode'] = 0;
     $result['memo'] = '缺少参数';
 }else if($details === 'null'){
-    $sql = "delete from cashier where order_id='$id'";
+    $sql = "delete from cashier where (user_id=$user_id and order_id='$id')";
     $data = $db->query($sql);
     if($data){
         $result['bizCode'] = 1;
@@ -27,10 +38,10 @@ if(!$id || !$details){
         $result['memo'] = '删除成功';
     }
 }else{
-    $sql = "update cashier set detail='$details' where order_id='$id'";
+    $sql = "update cashier set detail='$details' where (user_id=$user_id and order_id='$id')";
     $data = $db->query($sql);
     if($data){
-        $last_data = $db->queryUniqueObject("select * from cashier where order_id='$id'");
+        $last_data = $db->queryUniqueObject("select * from cashier where (user_id=$user_id and order_id='$id')");
         $result['bizCode'] = 1;
         $result['deleted'] = false;
         $result['data'] = array('sold'=>$last_data);
